+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### New Functions 🚀
 
-- `use_web_notification` - thanks for the help @centershocks44
+- `use_web_notification` - @centershocks44
 - `use_infinite_scroll`
+- `use_service_worker` - @lpotthast
 
 ### Breaking Changes 🛠