+++ CHANGELOG.md
@@ -3,13 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 
+## [0.8.0] - 2023-10-24
 
 ### New Functions 🚀
 
-- `use_web_notification` - @centershocks44
+- `use_web_notification` (thanks to @centershocks44)
 - `use_infinite_scroll`
-- `use_service_worker` - @lpotthast
+- `use_service_worker` (thanks to @lpotthast)
 
 ### Breaking Changes 🛠